Add ionCube loader php extension
As this is a zend extension it needs to be loaded manually rather than using the docker-php-ext-install command.

And since Docker doesn't support dynamic builds I have had to modify each 5.x version of php to include two new environment variables for the major and minor version of php being used.